updated qa home to be reuseable
