+++ src/additional-functions/workspace-section-and-store-names.js
@@ -1,6 +1,6 @@
 export const MONITORING_PLAN_STORE_NAME = "monitoringPlans";
 export const QA_CERT_TEST_SUMMARY_STORE_NAME = "qaCertTestSummary";
-
+export const EXPORT_STORE_NAME = "export";
 export const convertSectionToStoreName = (dataTableName) => {
   let storeName = "";
   switch (dataTableName) {
@@ -9,6 +9,9 @@ export const convertSectionToStoreName = (dataTableName) => {
       break;
     case QA_CERT_TEST_SUMMARY_STORE_NAME:
       storeName = QA_CERT_TEST_SUMMARY_STORE_NAME;
+      break;
+    case EXPORT_STORE_NAME:
+      storeName = EXPORT_STORE_NAME;
       break;
     default:
       break;
