+++ src/utils/constants/moduleTitles.js
@@ -1,0 +1,2 @@
+export const qa_Certifications_Test_Summary_Module = "QA Certifications Test Data "
+export const export_Module = "Export Data"
